requirements.txt
python-dotenv
azure-identity
uvicorn
azure-ai-projects
mcp
streamlit
streamlit-chat